% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_prism.R
\name{model_run}
\alias{model_run}
\title{Run codexcopd Model}
\usage{
model_run(model_input = NULL)
}
\arguments{
\item{model_input}{A list/json object with "age","charlson","FEV1",
"mMRC" (which is a number for the dyspnea scale of modified Medical Research Council)
,and "exacerbation"}
}
\value{
Return a dataset with four columns: CI_score, FEV1_score, MRC_score, exac_score
}
\description{
This function calls the codexcopd model.
}
